+++ packages/redux-devtools-inspector-monitor-test-tab/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 1.0.0
+
+### Minor Changes
+
+- 8a7eae4: Add React 18 to peerDependencies range
+
+### Patch Changes
+
+- Updated dependencies [8a7eae4]
+  - @redux-devtools/inspector-monitor@3.0.0
+  - @redux-devtools/ui@1.3.0
+
 ## 0.8.6
 
 ### Patch Changes