+++ packages/redux-devtools-inspector-monitor-trace-tab/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## 1.0.0
+
+### Minor Changes
+
+- 8a7eae4: Add React 18 to peerDependencies range
+
+### Patch Changes
+
+- Updated dependencies [8a7eae4]
+  - @redux-devtools/inspector-monitor@3.0.0
+
 ## 0.3.4
 
 ### Patch Changes