+++ packages/redux-devtools-inspector-monitor/demo/CHANGELOG.md
@@ -1,5 +1,14 @@
 # inspector-demo
 
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies [8a7eae4]
+  - @redux-devtools/dock-monitor@3.0.0
+  - @redux-devtools/inspector-monitor@3.0.0
+  - @redux-devtools/core@3.13.0
+
 ## 0.1.6
 
 ### Patch Changes